Refactor hotspots and fix seed PR gate
